Moved Interface Declarations
Interface declarations were previously in
`src/component-library/components.d.ts`. This clutters the namespace
unnecessarily. At the same time, packing the parameters for a component
into a single object (which can then be described by an interface) is
required by storybook.

The compromise is to retain the interfaces but move them to the file
they are used in. This means that storybook is happy, and if the
interface is indeed needed (such as with the login page component), this
interface can then be imported.

Moved the interfaces `*Props` from components.d.ts only for the
components rewritten to use TailwindCSS.
