+++ src/component-library/components/Text/Headings/H1.tsx
@@ -1,3 +1,7 @@
+interface TextProps {
+    text: string
+}
+
 export const H1 = ({text}: TextProps) => {
     return (
         <h1 className="text-4xl font-extrabold leading-none mt-4">{text}</h1>
